+++ sdk/lib/_internal/js_shared/lib/js_types.dart
@@ -8,138 +8,70 @@
 ///
 /// For consistency, all of the web backends have a version of this library.
 ///
-/// For the time being, all JS types are erased to their respective Dart type at
-/// runtime e.g. [JSString] -> [String]. Eventually, when we have inline
-/// classes, we may choose to either:
-///
-/// 1. Use [Object] as the representation type.
-/// 2. Have some analog to dart2wasm's [JSValue] as the representation type in
-/// order to separate the Dart and JS type hierarchies at runtime.
-/// 3. Continue using the respective Dart type.
-///
-/// Note that we can't use [Interceptor] to do option #2. [Interceptor] is a
-/// supertype of types like [interceptors.JSString], but not a supertype of the
-/// core types like [String]. This becomes relevant when we use external APIs.
-/// External APIs get lowered to `js_util` calls, which cast the return value.
-/// If a function returns a JavaScript string, it gets reified as a Dart
-/// [String] for example. Then when we cast to [JSString] in `js_util`, we get
-/// a cast failure, as [String] is not a subtype of [Interceptor].
-///
-/// For specific details of the JS type hierarchy, please see
-/// `sdk/lib/js_interop/js_interop.dart`.
+/// **WARNING**: You should *not* rely on these runtime types. Not only is this
+/// library not guaranteed to be consistent across platforms, these types may
+/// change in the future.
 library _js_types;
 
-import 'dart:_js_annotations';
-import 'dart:_js_helper' show createObjectLiteral;
+import 'dart:_native_typed_data' as typed_data;
+import 'dart:_interceptors' as interceptors;
 
-@JS()
-@staticInterop
-class JSAny {
-  // Unnamed factory constructor so users can only implement JSAny.
-  external factory JSAny._();
-}
+typedef JSAnyRepType = Object;
 
-@JS()
-@staticInterop
-class JSObject implements JSAny {
-  /// Returns a new object literal.
-  factory JSObject() => createObjectLiteral<JSObject>();
-}
+typedef JSObjectRepType = interceptors.JSObject;
 
-@JS()
-@staticInterop
-class JSFunction implements JSObject {}
+// TODO(srujzs): The JS function types have to be typed as
+// `LegacyJavaScriptObject` for now until we reify JS functions as
+// `JavaScriptFunction` instead of `LegacyJavaScriptObject` in DDC. This will
+// happen with the new RTI. Note that we *cannot* make this `Function`, even
+// though all JS functions are Dart functions in our type system, because
+// `Function` is not <: `JSObject`. This subtyping relationship is required for
+// `dart:js_interop`'s extension types.
+typedef JSFunctionRepType = interceptors.LegacyJavaScriptObject;
 
-@JS()
-@staticInterop
-class JSExportedDartFunction implements JSFunction {}
+typedef JSExportedDartFunctionRepType = interceptors.LegacyJavaScriptObject;
 
-@JS('Array')
-@staticInterop
-class JSArray implements JSObject {
-  external factory JSArray();
-  external factory JSArray.withLength(int length);
-}
+typedef JSArrayRepType = interceptors.JSArray<Object?>;
 
-@JS()
-@staticInterop
-class JSBoxedDartObject implements JSObject {}
+typedef JSBoxedDartObjectRepType = interceptors.JSObject;
 
-@JS()
-@staticInterop
-class JSArrayBuffer implements JSObject {}
+typedef JSArrayBufferRepType = typed_data.NativeByteBuffer;
 
-@JS()
-@staticInterop
-class JSDataView implements JSObject {}
+typedef JSDataViewRepType = typed_data.NativeByteData;
 
-@JS()
-@staticInterop
-class JSTypedArray implements JSObject {}
+typedef JSTypedArrayRepType = typed_data.NativeTypedData;
 
-@JS()
-@staticInterop
-class JSInt8Array implements JSTypedArray {}
+typedef JSInt8ArrayRepType = typed_data.NativeInt8List;
 
-@JS()
-@staticInterop
-class JSUint8Array implements JSTypedArray {}
+typedef JSUint8ArrayRepType = typed_data.NativeUint8List;
 
-@JS()
-@staticInterop
-class JSUint8ClampedArray implements JSTypedArray {}
+typedef JSUint8ClampedArrayRepType = typed_data.NativeUint8ClampedList;
 
-@JS()
-@staticInterop
-class JSInt16Array implements JSTypedArray {}
+typedef JSInt16ArrayRepType = typed_data.NativeInt16List;
 
-@JS()
-@staticInterop
-class JSUint16Array implements JSTypedArray {}
+typedef JSUint16ArrayRepType = typed_data.NativeUint16List;
 
-@JS()
-@staticInterop
-class JSInt32Array implements JSTypedArray {}
+typedef JSInt32ArrayRepType = typed_data.NativeInt32List;
 
-@JS()
-@staticInterop
-class JSUint32Array implements JSTypedArray {}
+typedef JSUint32ArrayRepType = typed_data.NativeUint32List;
 
-@JS()
-@staticInterop
-class JSFloat32Array implements JSTypedArray {}
+typedef JSFloat32ArrayRepType = typed_data.NativeFloat32List;
 
-@JS()
-@staticInterop
-class JSFloat64Array implements JSTypedArray {}
+typedef JSFloat64ArrayRepType = typed_data.NativeFloat64List;
 
-@JS()
-@staticInterop
-class JSNumber implements JSAny {}
+typedef JSNumberRepType = double;
 
-@JS()
-@staticInterop
-class JSBoolean implements JSAny {}
+typedef JSBooleanRepType = bool;
 
-@JS()
-@staticInterop
-class JSString implements JSAny {}
+typedef JSStringRepType = String;
 
-@JS()
-@staticInterop
-class JSSymbol implements JSAny {}
+typedef JSPromiseRepType = interceptors.JSObject;
 
-@JS()
-@staticInterop
-class JSBigInt implements JSAny {}
+typedef JSSymbolRepType = interceptors.JavaScriptSymbol;
+
+typedef JSBigIntRepType = interceptors.JavaScriptBigInt;
 
 /// [JSVoid] is just a typedef for [void]. While we could just use
 /// `JSUndefined`, in the future we may be able to use this to elide `return`s
 /// in JS trampolines.
-typedef JSVoid = void;
-
-@JS('Promise')
-@staticInterop
-class JSPromise implements JSObject {
-  external factory JSPromise(JSFunction executor);
-}
+typedef JSVoidRepType = void;