+++ sdk/lib/js_interop/js_interop.dart
@@ -6,10 +6,10 @@
 /// interop with JS. The JS type hierarchy is modeled after the actual type
 /// hierarchy in JS, and not the Dart type hierarchy.
 ///
-/// Note: The JS types that are exposed through this library are currently
-/// wrapper types that are erased to their runtime types. The runtime types will
-/// differ based on the backend. In general, stick to using conversion functions
-/// that are exposed as extension methods e.g. 'toJS'.
+/// Note: The JS types defined in this library only provide static guarantees.
+/// The runtime types differ based on the backend, so rely on static
+/// functionality like the conversion functions e.g. `toJS` and not runtime
+/// mechanisms like type checks and casts.
 ///
 /// **WARNING**:
 /// This library is still a work in progress. As such, JS types, allowed syntax,
@@ -19,7 +19,7 @@
 /// {@category Web}
 library dart.js_interop;
 
-import 'dart:_js_types' as js_types;
+import 'dart:_js_types';
 import 'dart:js_interop_unsafe';
 import 'dart:typed_data';
 
@@ -57,91 +57,134 @@ class JS {
 /// These are meant to separate the Dart and JS type hierarchies statically.
 ///
 /// **WARNING**:
-/// For now, the runtime semantics between backends may differ and may not be
-/// intuitive e.g. casting to [JSString] may give you inconsistent and
-/// surprising results depending on the value. It is preferred to always use the
-/// conversion functions e.g. `toJS` and `toDart`. The only runtime semantics
-/// stability we can guarantee is if a value actually is the JS type you are
-/// type-checking with/casting to e.g. `obj as JSString` will continue to work
-/// if `obj` actually is a JavaScript string.
+/// The runtime semantics between backends differ and may not be intuitive e.g.
+/// casting to [JSString] may give you inconsistent and surprising results
+/// depending on the value and the backend. It is preferred to always use the
+/// conversion functions e.g. `toJS` and `toDart`. You should always use interop
+/// to type-check e.g. `typeofEquals` and `instanceOfString` instead of relying
+/// on `is` and `as`, as the latter is backend-dependent.
 
 /// The overall top type in the JS types hierarchy.
-typedef JSAny = js_types.JSAny;
+extension type JSAny._(JSAnyRepType _jsAny) implements Object {}
 
-/// The representation type of all JavaScript objects for inline classes,
-/// [JSObject] <: [JSAny].
+/// The representation type of all JavaScript objects for extension types.
 ///
 /// This is the supertype of all JS objects, but not other JS types, like
-/// primitives. This is the only allowed `on` type for inline classes written by
-/// users to model JS interop objects. See https://dart.dev/web/js-interop for
-/// more details on how to use JS interop.
-// TODO(srujzs): This class _must_ be sealed before we can make this library
-// public. Either use the CFE mechanisms that exist today, or use the Dart 3
-// sealed classes feature.
-// TODO(joshualitt): Do we need to seal any other JS types on JS backends? We
-// probably want to seal all JS types on Wasm backends.
-// TODO(joshualitt): Add a [JSObject] constructor.
-typedef JSObject = js_types.JSObject;
+/// primitives. See https://dart.dev/web/js-interop for more details on how to
+/// use JS interop.
+extension type JSObject._(JSObjectRepType _jsObject) implements JSAny {
+  /// Constructor to go from an object from previous interop, like the types
+  /// from `package:js` or `dart:html`, to [JSObject].
+  ///
+  /// This and the public representation field are intended to avoid users
+  /// having to cast to and from [JSObject].
+  JSObject.fromInteropObject(Object interopObject)
+      : _jsObject = interopObject as JSObjectRepType;
 
-/// The type of all JS functions, [JSFunction] <: [JSObject].
-typedef JSFunction = js_types.JSFunction;
+  /// Returns a new object literal.
+  JSObject() : _jsObject = _createObjectLiteral();
+}
+
+// TODO(srujzs): Move this to `JSObject` once we can patch extension type
+// members.
+external JSObjectRepType _createObjectLiteral();
+
+/// The type of all JS functions.
+extension type JSFunction._(JSFunctionRepType _jsFunction)
+    implements JSObject {}
 
 /// The type of all Dart functions adapted to be callable from JS. We only allow
-/// a subset of Dart functions to be callable from JS, [JSExportedDartFunction]
-/// <: [JSFunction].
+/// a subset of Dart functions to be callable from JS.
 // TODO(joshualitt): Detail exactly what are the requirements.
-typedef JSExportedDartFunction = js_types.JSExportedDartFunction;
+extension type JSExportedDartFunction._(
+        JSExportedDartFunctionRepType _jsExportedDartFunction)
+    implements JSFunction {}
 
-/// The type of JS promises and promise-like objects, [JSPromise] <: [JSObject].
-typedef JSPromise = js_types.JSPromise;
+/// The type of JS promises and promise-like objects.
+@JS('Promise')
+extension type JSPromise._(JSPromiseRepType _jsPromise) implements JSObject {
+  external JSPromise(JSFunction executor);
+}
 
-/// The type of all JS arrays, [JSArray] <: [JSObject].
-typedef JSArray = js_types.JSArray;
+/// The type of all JS arrays.
+@JS('Array')
+extension type JSArray._(JSArrayRepType _jsArray) implements JSObject {
+  external JSArray();
+  external JSArray.withLength(int length);
+}
 
 /// The type of the boxed Dart object that can be passed to JS safely. There is
 /// no interface specified of this boxed object, and you may get a new box each
 /// time you box the same Dart object.
-/// [JSBoxedDartObject] <: [JSObject].
-typedef JSBoxedDartObject = js_types.JSBoxedDartObject;
+extension type JSBoxedDartObject._(JSBoxedDartObjectRepType _jsBoxedDartObject)
+    implements JSObject {}
 
-/// The type of JS array buffers, [JSArrayBuffer] <: [JSObject].
-typedef JSArrayBuffer = js_types.JSArrayBuffer;
+/// The type of JS' `ArrayBuffer`.
+extension type JSArrayBuffer._(JSArrayBufferRepType _jsArrayBuffer)
+    implements JSObject {}
 
-/// The type of JS byte data, [JSDataView] <: [JSObject].
-typedef JSDataView = js_types.JSDataView;
+/// The type of JS' `DataView`.
+extension type JSDataView._(JSDataViewRepType _jsDataView)
+    implements JSObject {}
 
-/// The abstract supertype of all JS typed arrays, [JSTypedArray] <: [JSObject].
-typedef JSTypedArray = js_types.JSTypedArray;
+/// The abstract supertype of all JS typed arrays.
+extension type JSTypedArray._(JSTypedArrayRepType _jsTypedArray)
+    implements JSObject {}
 
-/// The typed arrays themselves, `*Array` <: [JSTypedArray].
-typedef JSInt8Array = js_types.JSInt8Array;
-typedef JSUint8Array = js_types.JSUint8Array;
-typedef JSUint8ClampedArray = js_types.JSUint8ClampedArray;
-typedef JSInt16Array = js_types.JSInt16Array;
-typedef JSUint16Array = js_types.JSUint16Array;
-typedef JSInt32Array = js_types.JSInt32Array;
-typedef JSUint32Array = js_types.JSUint32Array;
-typedef JSFloat32Array = js_types.JSFloat32Array;
-typedef JSFloat64Array = js_types.JSFloat64Array;
+/// The type of JS' `Int8Array`.
+extension type JSInt8Array._(JSInt8ArrayRepType _jsInt8Array)
+    implements JSTypedArray {}
+
+/// The type of JS' `Uint8Array`.
+extension type JSUint8Array._(JSUint8ArrayRepType _jsUint8Array)
+    implements JSTypedArray {}
+
+/// The type of JS' `Uint8ClampedArray`.
+extension type JSUint8ClampedArray._(
+    JSUint8ClampedArrayRepType _jsUint8ClampedArray) implements JSTypedArray {}
+
+/// The type of JS' `Int16Array`.
+extension type JSInt16Array._(JSInt16ArrayRepType _jsInt16Array)
+    implements JSTypedArray {}
+
+/// The type of JS' `Uint16Array`.
+extension type JSUint16Array._(JSUint16ArrayRepType _jsUint16Array)
+    implements JSTypedArray {}
+
+/// The type of JS' `Int32Array`.
+extension type JSInt32Array._(JSInt32ArrayRepType _jsInt32Array)
+    implements JSTypedArray {}
+
+/// The type of JS' `Uint32Array`.
+extension type JSUint32Array._(JSUint32ArrayRepType _jsUint32Array)
+    implements JSTypedArray {}
+
+/// The type of JS' `Float32Array`.
+extension type JSFloat32Array._(JSFloat32ArrayRepType _jsFloat32Array)
+    implements JSTypedArray {}
+
+/// The type of JS' `Float64Array`.
+extension type JSFloat64Array._(JSFloat64ArrayRepType _jsFloat64Array)
+    implements JSTypedArray {}
 
 // The various JS primitive types. Crucially, unlike the Dart type hierarchy,
 // none of these are subtypes of [JSObject], but rather they are logically
 // subtypes of [JSAny].
 
-/// The type of JS numbers, [JSNumber] <: [JSAny].
-typedef JSNumber = js_types.JSNumber;
+/// The type of JS numbers.
+extension type JSNumber._(JSNumberRepType _jsNumber) implements JSAny {}
 
-/// The type of JS booleans, [JSBoolean] <: [JSAny].
-typedef JSBoolean = js_types.JSBoolean;
+/// The type of JS booleans.
+extension type JSBoolean._(JSBooleanRepType _jsBoolean) implements JSAny {}
 
-/// The type of JS strings, [JSString] <: [JSAny].
-typedef JSString = js_types.JSString;
+/// The type of JS strings.
+extension type JSString._(JSStringRepType _jsString) implements JSAny {}
 
-/// The type of JS Symbols, [JSSymbol] <: [JSAny].
-typedef JSSymbol = js_types.JSSymbol;
+/// The type of JS `Symbol`s.
+extension type JSSymbol._(JSSymbolRepType _jsSymbol) implements JSAny {}
 
-/// The type of JS BigInts, [JSBigInt] <: [JSAny].
-typedef JSBigInt = js_types.JSBigInt;
+/// The type of JS `BigInt`s.
+extension type JSBigInt._(JSBigIntRepType _jsBigInt) implements JSAny {}
 
 /// A getter to retrieve the global context that is used in static interop
 /// lowering.
@@ -207,11 +250,13 @@ extension JSObjectUtilExtension on JSObject {
 
 /// The type of `JSUndefined` when returned from functions. Unlike pure JS,
 /// no actual object will be returned.
-typedef JSVoid = js_types.JSVoid;
+// TODO(srujzs): Should we just remove this? There are no performance costs from
+// using `void`, and we'll likely provide a different way to box `undefined`.
+typedef JSVoid = JSVoidRepType;
 
 // Extension members to support conversions between Dart types and JS types.
 // Not all Dart types can be converted to JS types and vice versa.
-// TODO(joshualitt): We might want to investigate using inline classes instead
+// TODO(joshualitt): We might want to investigate using extension types instead
 // of extension methods for these methods.
 
 /// [JSExportedDartFunction] <-> [Function]