+++ tests/lib/js/static_interop_test/extension_type/non_interop_extension_type_static_test.dart
@@ -79,6 +79,7 @@ extension type EExtensionType(EJSObject _) {}
 
 @JS()
 extension type EExtensionType2(EExtensionType _) {}
+
 @JS()
 extension type ENonInterop._(EObject _) {
 //             ^